+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased:
+### ADDED:
+- Banner store entity
+
 ### CHANGED:
 - Refactor banner grid blocks
 - Refactor banner group blocks